(PC-40924)[API] fix: allow float value for device info in native endpoints

+++ api/src/pcapi/routes/native/v1/serialization/common_models.py
@@ -40,4 +40,4 @@ class DeviceInfo(HttpBodyModel):
     # TODO remove the following fields in future api version
     font_scale: float | None = None
     resolution: str | None = None
-    screen_zoom_level: int | None = None
+    screen_zoom_level: float | None = None
